backend/models/ltx2_animation.py
#!/usr/bin/env python3
"""
LTX-2 Animation Model for Video Generation
Handles animation-style video generation
"""

import argparse
import os
import sys
import torch
import torchvision.transforms as transforms
from PIL import Image
import numpy as np
import logging
import json
from datetime import datetime

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def parse_args():
    parser = argparse.ArgumentParser(description='Generate animated videos using LTX-2 model')
    parser.add_argument('--prompt', type=str, required=True, help='Text prompt for video generation')
    parser.add_argument('--duration', type=int, default=300, help='Duration in seconds')
    parser.add_argument('--resolution', type=str, default='1080p', help='Resolution (e.g., 720p, 1080p)')
    parser.add_argument('--output_dir', type=str, required=True, help='Output directory')
    parser.add_argument('--gpu', type=bool, default=True, help='Use GPU acceleration')
    return parser.parse_args()

def load_ltx2_model(model_path=None):
    """
    Load the LTX-2 animation model
    """
    try:
        # Placeholder for LTX-2 model loading
        # In production, replace with actual LTX-2 model
        logger.info("Initializing LTX-2 animation pipeline...")
        
        # Simulate model loading
        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        logger.info(f"Using device: {device}")
        
        # Return a mock model object
        class MockModel:
            def __init__(self):
                self.device = device
                
            def generate(self, prompt, num_frames, height, width):                # Simulate video generation
                frames = []
                for i in range(num_frames):
                    # Create animated frame based on prompt
                    frame = self.create_animated_frame(prompt, i, height, width)
                    frames.append(frame)
                return frames
            
            def create_animated_frame(self, prompt, frame_idx, height, width):
                # Create an animated frame based on the prompt
                img_array = np.zeros((height, width, 3), dtype=np.uint8)
                
                # Add animation based on frame index
                offset = (frame_idx * 5) % width
                for y in range(height):
                    for x in range(width):
                        # Create animated patterns
                        r = (x + offset) % 255
                        g = (y + frame_idx) % 255
                        b = (x + y + frame_idx) % 255
                        img_array[y, x] = [r, g, b]
                
                return Image.fromarray(img_array)
        
        return MockModel()
    except Exception as e:
        logger.error(f"Error loading LTX-2 model: {str(e)}")
        raise

def preprocess_animation_prompt(prompt):
    """
    Preprocess the prompt for animation style
    """
    # Add animation-specific keywords
    animation_keywords = [
        "animated", "cartoon style", "2D animation", 
        "smooth transitions", "fluid motion", "stylized"
    ]
    
    enhanced_prompt = f"{prompt}, " + ", ".join(animation_keywords)
    return enhanced_prompt

def generate_animation_frames(model, prompt, num_frames, height, width):
    """
    Generate animation frames using LTX-2 model
    """
    try:
        logger.info(f"Generating {num_frames} animation frames...")
        frames = model.generate(prompt, num_frames, height, width)
        logger.info("Frame generation completed")        return frames
    except Exception as e:
        logger.error(f"Error generating animation frames: {str(e)}")
        raise

def save_animation_video(frames, output_path, fps=12):
    """
    Save animation frames to video file using Pillow
    In production, this would use FFmpeg
    """
    try:
        # Save first frame as reference
        frames[0].save(
            output_path,
            save_all=True,
            append_images=frames[1:],
            duration=1000//fps,  # Duration in milliseconds per frame
            loop=0
        )
        logger.info(f"Animation video saved to {output_path}")
    except Exception as e:
        logger.error(f"Error saving animation video: {str(e)}")
        raise

def main():
    args = parse_args()
    
    logger.info(f"Starting animation video generation with prompt: '{args.prompt}'")
    logger.info(f"Duration: {args.duration}s, Resolution: {args.resolution}")
    
    try:
        # Load model
        logger.info("Loading LTX-2 animation model...")
        model = load_ltx2_model()
        
        # Preprocess prompt
        enhanced_prompt = preprocess_animation_prompt(args.prompt)
        logger.info(f"Enhanced prompt: {enhanced_prompt}")
        
        # Determine resolution
        width, height = (1920, 1080) if args.resolution == "1080p" else (1280, 720)
        
        # Calculate number of frames based on duration
        fps = 12  # Higher FPS for smoother animation
        num_frames = int(args.duration * fps)
        logger.info(f"Generating {num_frames} frames at {fps}fps")
        
        # Generate animation frames
        logger.info("Generating animation frames...")
        frames = generate_animation_frames(model, enhanced_prompt, num_frames, height, width)        
        # Create output directory if it doesn't exist
        os.makedirs(args.output_dir, exist_ok=True)
        
        # Generate output filename
        output_filename = f"animation_{args.prompt.replace(' ', '_')[:50]}_{int(os.urandom(4).hex(), 16)}.mp4"
        output_path = os.path.join(args.output_dir, output_filename)
        
        # Save video
        logger.info("Saving animation video...")
        save_animation_video(frames, output_path, fps)
        
        # Create metadata
        metadata = {
            "model_used": "LTX-2 Animation",
            "prompt": args.prompt,
            "duration": args.duration,
            "resolution": args.resolution,
            "fps": fps,
            "frame_count": len(frames),
            "output_file": output_filename,
            "generated_at": datetime.utcnow().isoformat()
        }
        
        metadata_path = output_path.replace('.mp4', '_metadata.json')
        with open(metadata_path, 'w') as f:
            json.dump(metadata, f, indent=2)
        
        logger.info(f"Animation generation completed successfully!")
        logger.info(f"Output: {output_path}")
        logger.info(f"Metadata: {metadata_path}")
        
        # Print completion message for parent process
        print(f"SUCCESS: Animation generated at {output_path}")
        
    except Exception as e:
        logger.error(f"Animation generation failed: {str(e)}")
        print(f"ERROR: {str(e)}")
        sys.exit(1)

if __name__ == "__main__":
    main()
