backend/models/requirements.txt
torch>=1.13.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.16.0
safetensors>=0.3.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.6.0
requests>=2.28.0
scipy>=1.9.0
matplotlib>=3.5.0
pandas>=1.4.0
